@@ -1,56 +1,53 @@
 {
-    "name": "formrelat",
-    "version": "2.2.0",
-    "description": "An Electron application with React",
+    "name": "formulario-triagem",
+    "version": "2.5.0",
+    "description": "Formulário de triagem infantil para avaliações",
     "main": "./out/main/index.js",
-    "author": "example.com",
+    "author": "Pedro Laurenti",
     "homepage": "https://electron-vite.org",
     "scripts": {
         "format": "prettier --write .",
-        "lint": "eslint . --ext .js,.jsx,.cjs,.mjs,.ts,.tsx,.cts,.mts --fix",
-        "start": "electron-vite preview",
+        "lint": "eslint . --fix",
+        "clean": "rimraf dist && rimraf out",
+        "postinstall": "npm run clean && electron-builder install-app-deps",
+
         "dev": "electron-vite dev",
-        "build": "electron-vite build",
-        "postinstall": "electron-builder install-app-deps",
-        "build:unpack": "npm run build && electron-builder --dir",
-        "build:win": "npm run build && electron-builder --win",
-        "build:mac": "npm run build && electron-builder --mac",
-        "build:linux": "npm run build && electron-builder --linux"
+        "start": "electron-vite preview",
+        "build": "npm run clean && electron-vite build",
+        
+        "build:unpack": "npm run clean && npm run build && electron-builder --dir",
+        "build:win": "npm run clean && npm run build && electron-builder --win",
+        "build:mac": "npm run clean && npm run build && electron-builder --mac",
+        "build:linux": "npm run clean && npm run build && electron-builder --linux"
     },
     "dependencies": {
         "@ckeditor/ckeditor5-build-classic": "^41.4.2",
         "@ckeditor/ckeditor5-react": "^7.0.0",
-        "@electron-toolkit/preload": "^3.0.0",
         "@electron-toolkit/utils": "^3.0.0",
         "@react-input/mask": "^1.2.5",
-        "draft-js": "^0.11.7",
-        "html2canvas": "^1.4.1",
-        "jspdf": "^2.5.1",
+        "autoprefixer": "^10.4.19",
+        "npm": "^10.8.2",
+        "prop-types": "^15.8.1",
         "puppeteer": "^22.11.2",
-        "react-icons": "^5.2.1",
-        "react-input-mask": "^2.0.4",
         "react-router-dom": "^6.23.1",
-        "react-to-print": "^2.15.1"
+        "sass": "^1.77.8",
+        "tailwindcss": "^3.4.6",
+        "uglify-js": "^3.19.0",
+        "vite-plugin-compression": "^0.5.1"
     },
     "devDependencies": {
-        "@electron-toolkit/eslint-config": "^1.0.2",
-        "@electron-toolkit/eslint-config-prettier": "^2.0.0",
-        "@types/draft-js": "^0.11.18",
+        "@eslint/js": "^9.7.0",
+        "@rollup/plugin-terser": "^0.4.4",
         "@types/react": "^18.3.3",
-        "@types/react-input-mask": "^3.0.5",
         "@vitejs/plugin-react": "^4.2.1",
-        "autoprefixer": "^10.4.19",
         "electron": "^28.2.0",
         "electron-builder": "^24.9.1",
         "electron-vite": "^2.0.0",
-        "eslint": "^8.56.0",
-        "eslint-config-prettier": "^9.1.0",
-        "eslint-plugin-react": "^7.33.2",
+        "eslint": "^9.7.0",
+        "globals": "^15.8.0",
         "prettier": "^3.2.4",
         "react": "^18.3.1",
         "react-dom": "^18.3.1",
-        "sass": "^1.77.6",
-        "tailwindcss": "^3.4.4",
         "vite": "^5.0.12"
     }
 }
